test(pii): use real maskEmail in mocks via importActual
Self-review follow-up on PR #900. The PII scrub commit added inline
maskEmail re-implementations inside every vi.mock('@pagespace/lib/server')
factory. If the real maskEmail policy ever drifted (shorter mask,
domain stripping, etc.) the mocks would silently keep passing while
production diverged.

- Extract maskEmail to a zero-import leaf module (audit/mask-email.ts)
  with a doc comment explaining why the domain is intentionally retained
  (OAuth-provider debugging vs. local-part identification trade-off)
- Add ./audit/mask-email subpath export and vitest alias so tests can
  pull the real implementation via vi.importActual without dragging in
  the DB-bound audit service barrel
- Convert all 15 vi.mock('@pagespace/lib/server') factories to async
  form that imports the real maskEmail
- Drop the __GIFT_TEST_MODE__ env toggle and selectCount % 2 modulo
  trick from gift-subscription PII test; switch to per-test
  mockReturnValueOnce queue (vi.hoisted for TDZ safety)

All 16 PR-touched test files green (401 tests). Typecheck + lint clean.
